refactor: remove dummy search repository impl
